Add Transactions window

+++ cxx/ui/widget/filterinput.h
@@ -9,7 +9,7 @@
 class FilterInput : public QLineEdit
 {
     Q_OBJECT
-    static constexpr QChar icon = QChar{finances::FontIcon::Filter};
+    static constexpr QChar icon = QChar{uint(finances::FontIcon::Filter)};
 protected Q_SLOTS:
     void onTextChanged(const QString &text);
 public:
